Passing relevant event maps into DeviceConnection
